Fixed failing unit tests
